feat: Enable [un]installation and upgrade of multiple packages at once. Closes #11

+++ pkg/lib/commands/uninstall.sh
@@ -1,21 +1,27 @@
 # shellcheck shell=bash
 
 basher-uninstall() {
-	local package="$1"
-
-	local site= user= repository= ref=
-	util.parse_package_full "$1"
-	IFS=':' read -r site user repository ref <<< "$REPLY"
-
-	local package="$user/$repository"
-
-	if [ ! -d "$NEOBASHER_PACKAGES_PATH/$package" ]; then
-		die "Package '$package' is not installed"
+	if (( $# == 0 )); then
+		die "You must supply at least one package"
 	fi
 
-	basher-plumbing-unlink-man "$package"
-	basher-plumbing-unlink-bins "$package"
-	basher-plumbing-unlink-completions "$package"
+	for repoSpec; do
+		local site= user= repository= ref=
+		util.parse_package_full "$repoSpec"
+		IFS=':' read -r site user repository ref <<< "$REPLY"
 
-	rm -rf "${NEOBASHER_PACKAGES_PATH:?}/$package"
+		log.info "Uninstalling '$repoSpec'"
+
+		local package="$user/$repository"
+
+		if [ ! -d "$NEOBASHER_PACKAGES_PATH/$package" ]; then
+			die "Package '$package' is not installed"
+		fi
+
+		basher-plumbing-unlink-man "$package"
+		basher-plumbing-unlink-bins "$package"
+		basher-plumbing-unlink-completions "$package"
+
+		rm -rf "${NEOBASHER_PACKAGES_PATH:?}/$package"
+	done
 }
